Implement pointer validity constraint (#55)
Many of expressions generated by the fuzzer used to result in an "invalid memory access" due to dereferencing invalid pointers, which made focusing on relevant expressions harder. This change implements a constraint to check whether an invalid pointer can be generated.

The constraint is used in such way that a valid pointer is generated only by taking address of variables (or fields of variables) and is used without any offset. Expressions such as `*(ptr + offset)` or `arr[0]` aren't allowed, unless there is an address-of operator `&` in front of them.