Improve memory access constraints in fuzzer (#63)
This change tightens memory access control when creating expressions by keeping a track of number of times an expression can be dereferenced ("freedom index"). Before this change, a single bool was used to determine whether an expression must be a valid pointer, but that lead to issues with expressions such as `**ptr`, where `ptr` is a valid pointer that can only be dereferenced once.

In addition, a small change to comparing LLDB and lldb-eval types is introduced that ignores type qualifiers to reduce unwanted noise produced when running the fuzzer.